refactor(cef): split client.rs into mod + handlers + wndproc + helpers (task #182 PR-G)
Splits client.rs (1942 lines) into 4 sibling files. Pure mechanical
extraction of structurally-cohesive sections that don't reference the
big AgentMuxHandler impl block. The big methods (on_after_created,
on_before_close, on_load_*) stay in mod.rs because their bodies access
many private `self` fields — extracting them requires real refactoring
beyond the mechanical scope of this PR.

## Final layout

| File | Lines | Contents |
|---|---|---|
| client/mod.rs | 1304 | imports, AgentMuxHandler struct + impl (the big one), pub re-exports |
| client/handlers.rs | 324 | Cef handler trait wrappers (wrap_client!, wrap_display_handler!, wrap_drag_handler!, wrap_focus_handler!, wrap_keyboard_handler!, wrap_life_span_handler!, wrap_load_handler!, wrap_request_handler!) |
| client/wndproc.rs | 250 | Win32 frameless window setup (setup_native_frameless, install_frameless_resize_hook, set_window_icon, skip_taskbar) |
| client/helpers.rs | 104 | js_string_literal, html_escape, backend_close_window |

mod.rs went 1942 → 1304 lines (-638). Each sibling file is focused on
a single layer (handler glue / Win32 / utils).

## Cross-module wiring

- handlers.rs: imports AgentMuxHandler + dlog from super; the wrap_*
  macros come from cef::*.
- wndproc.rs: extracted fns marked pub(super); skip_taskbar called
  from mod.rs's on_after_created.
- helpers.rs: extracted fns marked pub(super); imports dlog from super.
- mod.rs: brings everything back via use statements at the top of the
  impl block.

## NOT done (out of mechanical scope)

The two giant impl methods (on_after_created ~308 lines, on_before_close
~398 lines) stay in mod.rs. Splitting them requires either:
1. Extracting their bodies into free fns that take handler-state params
   (changes the call shape), or
2. Splitting into trait-impl modules with shared private state
   (substantial refactor)

Both are beyond mechanical scope. If desired, follow-up PR can do
this — the savings would be ~700 more lines from mod.rs.

63 host tests pass.

## Status of task #182

All planned splits done:
- PR-A through PR-E: launcher reducer (5 PRs)
- PR-F-1, PR-F-2: host reducer (2 PRs)
- PR-G: client.rs (this PR)
