+++ agentmux-cef/src/client/mod.rs
@@ -78,6 +78,17 @@ pub struct AgentMuxHandler {
     is_browser_pane: bool,
 }
 
+mod handlers;
+mod helpers;
+#[cfg(target_os = "windows")]
+mod wndproc;
+
+pub use handlers::AgentMuxClient;
+
+use helpers::{js_string_literal, html_escape, backend_close_window};
+#[cfg(target_os = "windows")]
+use wndproc::{install_frameless_resize_hook, set_window_icon, setup_native_frameless, skip_taskbar};
+
 impl AgentMuxHandler {
     pub fn new(state: Arc<AppState>, ipc_port: u16) -> Arc<Mutex<Self>> {
         Self::new_with_browser_pane(state, ipc_port, false)
@@ -1286,657 +1297,8 @@ impl AgentMuxHandler {
     }
 }
 
-/// Quote a string as a JavaScript string literal — escape backslashes,
-/// quotes, and newlines so it's safe to embed inside `<script>` via
-/// `format!`. Used by the recovery page to inject the app URL for the
-/// Reload button's navigation target.
-fn js_string_literal(s: &str) -> String {
-    let mut out = String::with_capacity(s.len() + 2);
-    out.push('"');
-    for c in s.chars() {
-        match c {
-            '\\' => out.push_str("\\\\"),
-            '"' => out.push_str("\\\""),
-            '\n' => out.push_str("\\n"),
-            '\r' => out.push_str("\\r"),
-            '\t' => out.push_str("\\t"),
-            '<' => out.push_str("\\u003c"), // defense against </script> injection
-            '>' => out.push_str("\\u003e"),
-            '&' => out.push_str("\\u0026"),
-            c if (c as u32) < 0x20 => {
-                out.push_str(&format!("\\u{:04x}", c as u32));
-            }
-            c => out.push(c),
-        }
-    }
-    out.push('"');
-    out
-}
-
-/// Minimal HTML escape for the recovery page. Only the characters that
-/// would break the `format!`-templated string need attention; the input
-/// (CEF status enum + cef-provided error string) is trusted but may
-/// contain `&` / `<` / `>` in some failure modes.
-fn html_escape(s: &str) -> String {
-    s.replace('&', "&amp;")
-        .replace('<', "&lt;")
-        .replace('>', "&gt;")
-        .replace('"', "&quot;")
-        .replace('\'', "&#39;")
-}
 
 // ---------------------------------------------------------------------------
 // CefClient — routes to sub-handlers
-// ---------------------------------------------------------------------------
 
-wrap_client! {
-    pub struct AgentMuxClient {
-        inner: Arc<Mutex<AgentMuxHandler>>,
-        is_browser_pane: bool,
-    }
 
-    impl Client {
-        fn display_handler(&self) -> Option<DisplayHandler> {
-            Some(AgentMuxDisplayHandler::new(self.inner.clone()))
-        }
-
-        fn keyboard_handler(&self) -> Option<KeyboardHandler> {
-            Some(AgentMuxKeyboardHandler::new())
-        }
-
-        fn life_span_handler(&self) -> Option<LifeSpanHandler> {
-            Some(AgentMuxLifeSpanHandler::new(self.inner.clone()))
-        }
-
-        fn load_handler(&self) -> Option<LoadHandler> {
-            Some(AgentMuxLoadHandler::new(self.inner.clone()))
-        }
-
-        fn request_handler(&self) -> Option<RequestHandler> {
-            Some(AgentMuxRequestHandler::new(self.inner.clone()))
-        }
-
-        fn drag_handler(&self) -> Option<DragHandler> {
-            if self.is_browser_pane {
-                return None;
-            }
-            Some(AgentMuxDragHandler::new(self.inner.clone()))
-        }
-
-        fn focus_handler(&self) -> Option<FocusHandler> {
-            // For browser panes only: cancel CEF's auto-focus on navigation so the
-            // child HWND doesn't steal keyboard focus from the main window when the
-            // page finishes loading. The user can still click into the pane to focus it.
-            if self.is_browser_pane {
-                Some(AgentMuxPaneFocusHandler::new())
-            } else {
-                None
-            }
-        }
-    }
-}
-
-// FocusHandler used only by browser-pane clients. Returns 0 for every
-// focus source (never cancels at the CEF level) — cancelling NAVIGATION
-// focus during the very first navigation of a newly-created pane fires
-// CEF's `on_before_close` on that pane ~10ms later. Focus-steal
-// protection lives entirely in the Win32 `WndProc` subclass below
-// (`browser_pane::hwnd::install_browser_pane_focus_redirect`), which redirects programmatic
-// `WM_SETFOCUS` back to the top-level window. User clicks are let through
-// because `WM_LBUTTONDOWN` in the subclass arms `ALLOW_BROWSER_PANE_FOCUS_ONCE`.
-wrap_focus_handler! {
-    struct AgentMuxPaneFocusHandler;
-
-    impl FocusHandler {
-        fn on_set_focus(
-            &self,
-            _browser: Option<&mut Browser>,
-            source: FocusSource,
-        ) -> ::std::os::raw::c_int {
-            // Previously we cancelled FocusSource::NAVIGATION here to
-            // stop page-load from stealing focus away from the main
-            // window. But cancelling on_set_focus during the very
-            // first navigation of a newly-created pane triggered CEF
-            // to fire `on_before_close` on that pane ~10ms later —
-            // reliably reproducible when creating a 2nd browser pane.
-            // The Win32 WndProc subclass below already redirects
-            // page-load SetFocus to the top-level window (see
-            // `browser_pane::hwnd::install_browser_pane_focus_redirect`), which
-            // handles the original focus-steal concern. Returning 0
-            // here so CEF proceeds with normal focus handling at the
-            // Chromium level; Win32 subclass continues to redirect
-            // any resulting Win32 focus change away from the pane.
-            tracing::info!("[pane-focus] on_set_focus source={:?} cancel=false", source);
-            0
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// DragHandler — handles `-webkit-app-region: drag` regions reported by the
-// renderer (used on macOS/Windows where native draggable regions work).
-//
-// NOTE(Linux): On Linux/Wayland we do NOT use -webkit-app-region: drag for
-// window-move because Chromium suppresses ALL events on drag regions before
-// they reach the renderer (verified empirically), making drag mutually
-// exclusive with right-click contextmenu on the same element. Linux drag is
-// JS-driven instead — see frontend/app/hook/useWindowDrag.linux.ts and
-// the start_window_drag IPC → CefWindow::BeginWindowDrag() (CEF source
-// patch in agentmux/7680-... branch). Retro:
-// docs/retros/2026-05-02-drag-and-rightclick-coexistence.md.
-
-wrap_drag_handler! {
-    struct AgentMuxDragHandler {
-        inner: Arc<Mutex<AgentMuxHandler>>,
-    }
-
-    impl DragHandler {
-        fn on_draggable_regions_changed(
-            &self,
-            browser: Option<&mut Browser>,
-            _frame: Option<&mut Frame>,
-            regions: Option<&[DraggableRegion]>,
-        ) {
-            if let Some(rs) = regions {
-                let summary: Vec<String> = rs.iter().map(|r| {
-                    format!("{}x{}@{},{} drag={}", r.bounds.width, r.bounds.height, r.bounds.x, r.bounds.y, r.draggable != 0)
-                }).collect();
-                tracing::info!("[drag_handler] on_draggable_regions_changed: {} regions — {:?}", rs.len(), summary);
-            } else {
-                tracing::info!("[drag_handler] on_draggable_regions_changed: None");
-            }
-            let mut browser = browser.cloned();
-            let Some(browser_view) = browser_view_get_for_browser(browser.as_mut()) else { return };
-            let Some(window) = browser_view.window() else { return };
-            window.set_draggable_regions(regions);
-        }
-    }
-}
-
-// KeyboardHandler — intercept Ctrl+<key> shortcuts before CEF/Chromium
-// consumes them (e.g., Ctrl+P = print, Ctrl+G = find-next).
-// Returning true from on_pre_key_event tells CEF "handled" so it won't
-// trigger the built-in action; the key still reaches JavaScript.
-// ---------------------------------------------------------------------------
-
-/// CEF event flag: Ctrl key is held.
-const EVENTFLAG_CONTROL_DOWN: u32 = 1 << 2;
-
-/// Windows virtual-key codes for shortcuts we want to forward to JS.
-const VK_P: i32 = 0x50; // Ctrl+P — command palette (not print)
-const VK_G: i32 = 0x47; // Ctrl+G — (reserve for app use)
-
-wrap_keyboard_handler! {
-    struct AgentMuxKeyboardHandler;
-
-    impl KeyboardHandler {
-        fn on_pre_key_event(
-            &self,
-            _browser: Option<&mut Browser>,
-            event: Option<&KeyEvent>,
-            _os_event: Option<&mut crate::OsKeyEvent>,
-            is_keyboard_shortcut: Option<&mut ::std::os::raw::c_int>,
-        ) -> ::std::os::raw::c_int {
-            if let Some(ev) = event {
-                let ctrl = (ev.modifiers & EVENTFLAG_CONTROL_DOWN) != 0;
-                if ctrl && matches!(ev.windows_key_code, VK_P | VK_G) {
-                    // Tell CEF this is a keyboard shortcut so it dispatches
-                    // the keydown event to JavaScript instead of handling it
-                    // as a built-in browser action (print dialog, etc.).
-                    if let Some(flag) = is_keyboard_shortcut {
-                        *flag = 1;
-                    }
-                    // Return 0 = not consumed at pre-key stage; CEF will
-                    // still call on_key_event where we return 0 again,
-                    // letting JS handle it via the normal keydown path.
-                }
-            }
-            0 // not consumed
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// DisplayHandler — title changes
-// ---------------------------------------------------------------------------
-
-wrap_display_handler! {
-    struct AgentMuxDisplayHandler {
-        inner: Arc<Mutex<AgentMuxHandler>>,
-    }
-
-    impl DisplayHandler {
-        fn on_title_change(&self, browser: Option<&mut Browser>, title: Option<&CefString>) {
-            let mut inner = self.inner.lock();
-            inner.on_title_change(browser, title);
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// LifeSpanHandler — browser creation/destruction
-// ---------------------------------------------------------------------------
-
-wrap_life_span_handler! {
-    struct AgentMuxLifeSpanHandler {
-        inner: Arc<Mutex<AgentMuxHandler>>,
-    }
-
-    impl LifeSpanHandler {
-        fn on_after_created(&self, browser: Option<&mut Browser>) {
-            let mut inner = self.inner.lock();
-            inner.on_after_created(browser);
-        }
-
-        fn do_close(&self, browser: Option<&mut Browser>) -> i32 {
-            let mut inner = self.inner.lock();
-            inner.do_close(browser).into()
-        }
-
-        fn on_before_close(&self, browser: Option<&mut Browser>) {
-            let mut inner = self.inner.lock();
-            inner.on_before_close(browser);
-        }
-
-        fn on_before_popup(
-            &self,
-            browser: Option<&mut Browser>,
-            frame: Option<&mut Frame>,
-            _popup_id: ::std::os::raw::c_int,
-            target_url: Option<&CefString>,
-            _target_frame_name: Option<&CefString>,
-            target_disposition: WindowOpenDisposition,
-            _user_gesture: ::std::os::raw::c_int,
-            _popup_features: Option<&PopupFeatures>,
-            _window_info: Option<&mut WindowInfo>,
-            _client: Option<&mut Option<Client>>,
-            _settings: Option<&mut BrowserSettings>,
-            _extra_info: Option<&mut Option<DictionaryValue>>,
-            _no_javascript_access: Option<&mut ::std::os::raw::c_int>,
-        ) -> ::std::os::raw::c_int {
-            let mut inner = self.inner.lock();
-            if inner.on_before_popup(browser, frame, target_url, target_disposition) {
-                1
-            } else {
-                0
-            }
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// LoadHandler — load events and errors
-// ---------------------------------------------------------------------------
-
-wrap_load_handler! {
-    struct AgentMuxLoadHandler {
-        inner: Arc<Mutex<AgentMuxHandler>>,
-    }
-
-    impl LoadHandler {
-        fn on_loading_state_change(
-            &self,
-            browser: Option<&mut Browser>,
-            is_loading: ::std::os::raw::c_int,
-            can_go_back: ::std::os::raw::c_int,
-            can_go_forward: ::std::os::raw::c_int,
-        ) {
-            let mut inner = self.inner.lock();
-            inner.on_loading_state_change(browser, is_loading, can_go_back, can_go_forward);
-        }
-
-        fn on_load_end(
-            &self,
-            browser: Option<&mut Browser>,
-            frame: Option<&mut Frame>,
-            http_status_code: i32,
-        ) {
-            let mut inner = self.inner.lock();
-            inner.on_load_end(browser, frame, http_status_code);
-        }
-
-        fn on_load_error(
-            &self,
-            browser: Option<&mut Browser>,
-            frame: Option<&mut Frame>,
-            error_code: Errorcode,
-            error_text: Option<&CefString>,
-            failed_url: Option<&CefString>,
-        ) {
-            let mut inner = self.inner.lock();
-            inner.on_load_error(browser, frame, error_code, error_text, failed_url);
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// RequestHandler — render-process termination (white-screen recovery)
-// ---------------------------------------------------------------------------
-//
-// We only override `on_render_process_terminated` here. Everything else
-// inherits the default (no-op) implementations from the cef-rs trait.
-// See SPEC_GRACEFUL_CRASH_HANDLING_2026_04_13.md (PR 1).
-
-wrap_request_handler! {
-    struct AgentMuxRequestHandler {
-        inner: Arc<Mutex<AgentMuxHandler>>,
-    }
-
-    impl RequestHandler {
-        fn on_render_process_terminated(
-            &self,
-            browser: Option<&mut Browser>,
-            status: TerminationStatus,
-            error_code: ::std::os::raw::c_int,
-            error_string: Option<&CefString>,
-        ) {
-            let mut inner = self.inner.lock();
-            inner.on_render_process_terminated(browser, status, error_code, error_string);
-        }
-    }
-}
-
-/// Set up a native frameless window: extend client area over the thick frame
-/// border so the resize handle is invisible, then subclass the window to
-/// handle WM_NCHITTEST for edge resize.
-///
-/// DwmExtendFrameIntoClientArea(-1) makes the entire frame transparent, but
-/// it also removes the non-client hit-test region. Without the subclass,
-/// Windows can't tell which part of the window edge should be a resize handle.
-/// The subclass returns HT{LEFT,RIGHT,TOP,BOTTOM,TOPLEFT,...} when the cursor
-/// is within RESIZE_BORDER pixels of the window edge.
-#[cfg(target_os = "windows")]
-unsafe fn setup_native_frameless(hwnd: *mut std::ffi::c_void) {
-    use windows_sys::Win32::Graphics::Dwm::DwmExtendFrameIntoClientArea;
-    use windows_sys::Win32::UI::Controls::MARGINS;
-
-    let margins = MARGINS {
-        cxLeftWidth: -1,
-        cxRightWidth: -1,
-        cyTopHeight: -1,
-        cyBottomHeight: -1,
-    };
-    let result = DwmExtendFrameIntoClientArea(hwnd, &margins);
-    if result == 0 {
-        tracing::info!("Applied DwmExtendFrameIntoClientArea to hide resize border");
-    } else {
-        tracing::warn!("DwmExtendFrameIntoClientArea failed: hr={:#x}", result);
-    }
-}
-
-/// Map of HWND -> original WndProc for secondary windows with edge resize hooks.
-/// Stored here instead of GWLP_USERDATA to avoid clobbering CEF's data.
-#[cfg(target_os = "windows")]
-static ORIGINAL_WNDPROCS: std::sync::LazyLock<
-    std::sync::Mutex<std::collections::HashMap<usize, isize>>,
-> = std::sync::LazyLock::new(|| std::sync::Mutex::new(std::collections::HashMap::new()));
-
-// Pane Win32 focus-redirect subclass + ALLOW_BROWSER_PANE_FOCUS_ONCE flag moved to
-// `crate::browser_pane::hwnd` in Phase 2 of the modularization split. See
-// `docs/specs/SPEC_BROWSER_PANE_MODULARIZATION.md`.
-
-/// Install a WndProc hook on a SECONDARY window that handles:
-/// - WM_NCCALCSIZE: returns 0 to eliminate the non-client area (removes the
-///   wide title bar / top border that WS_THICKFRAME + DWM extension creates)
-/// - WM_NCHITTEST: returns HT{LEFT,RIGHT,...} for resize zones at window edges
-///
-/// MUST NOT be installed on the main CEF Views window — that window handles
-/// resize through its delegate, and hooking it clobbers CEF internals.
-#[cfg(target_os = "windows")]
-unsafe fn install_frameless_resize_hook(hwnd: *mut std::ffi::c_void) {
-    use windows_sys::Win32::UI::WindowsAndMessaging::*;
-
-    const RESIZE_BORDER: i32 = 6;
-
-    unsafe extern "system" fn wndproc_hook(
-        hwnd: *mut std::ffi::c_void,
-        msg: u32,
-        wparam: usize,
-        lparam: isize,
-    ) -> isize {
-        match msg {
-            // Remove the non-client area entirely — this eliminates the wide
-            // top border that WS_THICKFRAME normally reserves for the title bar.
-            WM_NCCALCSIZE if wparam == 1 => {
-                // Returning 0 with wparam=1 tells Windows the client area
-                // fills the entire window rect. No title bar, no borders.
-                return 0;
-            }
-
-            // Suppress the DWM activation border — return TRUE without
-            // calling DefWindowProc so Windows doesn't repaint the frame.
-            WM_NCACTIVATE => {
-                return 1; // TRUE = allow activation, but skip default border paint
-            }
-
-            WM_NCHITTEST => {
-                let x = (lparam & 0xFFFF) as i16 as i32;
-                let y = ((lparam >> 16) & 0xFFFF) as i16 as i32;
-
-                let mut rect = std::mem::zeroed::<windows_sys::Win32::Foundation::RECT>();
-                GetWindowRect(hwnd, &mut rect);
-
-                let left = x - rect.left < RESIZE_BORDER;
-                let right = rect.right - x < RESIZE_BORDER;
-                let top = y - rect.top < RESIZE_BORDER;
-                let bottom = rect.bottom - y < RESIZE_BORDER;
-
-                if top && left { return HTTOPLEFT as isize; }
-                if top && right { return HTTOPRIGHT as isize; }
-                if bottom && left { return HTBOTTOMLEFT as isize; }
-                if bottom && right { return HTBOTTOMRIGHT as isize; }
-                if left { return HTLEFT as isize; }
-                if right { return HTRIGHT as isize; }
-                if top { return HTTOP as isize; }
-                if bottom { return HTBOTTOM as isize; }
-                // Not on an edge — fall through to original WndProc.
-            }
-
-            _ => {}
-        }
-
-        // Delegate to the original WndProc.
-        let key = hwnd as usize;
-        let original = ORIGINAL_WNDPROCS
-            .lock()
-            .ok()
-            .and_then(|map| map.get(&key).copied())
-            .unwrap_or(0);
-        if original != 0 {
-            CallWindowProcW(Some(std::mem::transmute(original)), hwnd, msg, wparam, lparam)
-        } else {
-            DefWindowProcW(hwnd, msg, wparam, lparam)
-        }
-    }
-
-    let original = GetWindowLongPtrW(hwnd, GWLP_WNDPROC);
-    if let Ok(mut map) = ORIGINAL_WNDPROCS.lock() {
-        map.insert(hwnd as usize, original);
-    }
-    SetWindowLongPtrW(hwnd, GWLP_WNDPROC, wndproc_hook as isize);
-    tracing::info!("Installed frameless resize hook (WM_NCCALCSIZE + WM_NCHITTEST)");
-}
-
-/// Hide the given top-level HWND from the Windows taskbar via
-/// `ITaskbarList::DeleteTab`. The window remains fully usable — Alt-Tab still
-/// finds it, it takes focus, repaints, etc. — but the shell paints no taskbar
-/// button for it regardless of the user's "Combine taskbar buttons" setting.
-///
-/// Used only for `WindowKind::Subwindow` top-level windows. Must be called
-/// once the HWND exists (post-`on_after_created`) and re-applied on the
-/// `TaskbarCreated` broadcast after Explorer restarts.
-///
-/// Same primitive Electron uses in `NativeWindowViews::SetSkipTaskbar`
-/// (`shell/browser/native_window_views.cc`).
-#[cfg(target_os = "windows")]
-unsafe fn skip_taskbar(hwnd: *mut std::ffi::c_void) {
-    use windows_sys::Win32::System::Com::{CoCreateInstance, CLSCTX_INPROC_SERVER};
-    use windows_sys::core::GUID;
-
-    // CLSID_TaskbarList
-    const CLSID_TASKBAR_LIST: GUID = GUID {
-        data1: 0x56FDF344,
-        data2: 0xFD6D,
-        data3: 0x11D0,
-        data4: [0x95, 0x8A, 0x00, 0x60, 0x97, 0xC9, 0xA0, 0x90],
-    };
-    // IID_ITaskbarList
-    const IID_TASKBAR_LIST: GUID = GUID {
-        data1: 0x56FDF342,
-        data2: 0xFD6D,
-        data3: 0x11D0,
-        data4: [0x95, 0x8A, 0x00, 0x60, 0x97, 0xC9, 0xA0, 0x90],
-    };
-
-    // Hand-rolled vtable — `windows-sys` doesn't expose `ITaskbarList` types
-    // at this feature level, and pulling in the full `windows` crate for one
-    // COM interface is overkill.
-    #[repr(C)]
-    struct ITaskbarList {
-        lp_vtbl: *const ITaskbarListVtbl,
-    }
-    #[repr(C)]
-    struct ITaskbarListVtbl {
-        query_interface: unsafe extern "system" fn(*mut ITaskbarList, *const GUID, *mut *mut core::ffi::c_void) -> i32,
-        add_ref: unsafe extern "system" fn(*mut ITaskbarList) -> u32,
-        release: unsafe extern "system" fn(*mut ITaskbarList) -> u32,
-        hr_init: unsafe extern "system" fn(*mut ITaskbarList) -> i32,
-        add_tab: unsafe extern "system" fn(*mut ITaskbarList, *mut core::ffi::c_void) -> i32,
-        delete_tab: unsafe extern "system" fn(*mut ITaskbarList, *mut core::ffi::c_void) -> i32,
-        activate_tab: unsafe extern "system" fn(*mut ITaskbarList, *mut core::ffi::c_void) -> i32,
-        set_active_alt: unsafe extern "system" fn(*mut ITaskbarList, *mut core::ffi::c_void) -> i32,
-    }
-
-    let mut tbl: *mut ITaskbarList = std::ptr::null_mut();
-    let hr = CoCreateInstance(
-        &CLSID_TASKBAR_LIST as *const GUID,
-        std::ptr::null_mut(),
-        CLSCTX_INPROC_SERVER,
-        &IID_TASKBAR_LIST as *const GUID,
-        &mut tbl as *mut _ as *mut _,
-    );
-    if hr < 0 || tbl.is_null() {
-        tracing::warn!("[skip_taskbar] CoCreateInstance(TaskbarList) failed: hr=0x{:x}", hr);
-        return;
-    }
-
-    let vtbl = &*(*tbl).lp_vtbl;
-    let hr = (vtbl.hr_init)(tbl);
-    if hr < 0 {
-        tracing::warn!("[skip_taskbar] HrInit failed: hr=0x{:x}", hr);
-        (vtbl.release)(tbl);
-        return;
-    }
-    let hr = (vtbl.delete_tab)(tbl, hwnd);
-    if hr < 0 {
-        tracing::warn!("[skip_taskbar] DeleteTab failed: hr=0x{:x}", hr);
-    } else {
-        tracing::info!("[skip_taskbar] hid HWND {:p} from taskbar", hwnd);
-    }
-    (vtbl.release)(tbl);
-}
-
-/// Load the app icon from the exe's embedded resource and set it on the window.
-/// This makes the icon appear in the taskbar and Alt+Tab switcher instead of
-/// the default CEF/Chromium icon.
-#[cfg(target_os = "windows")]
-unsafe fn set_window_icon(hwnd: *mut std::ffi::c_void) {
-    use windows_sys::Win32::UI::WindowsAndMessaging::*;
-    use windows_sys::Win32::System::LibraryLoader::GetModuleHandleW;
-
-    let hinstance = GetModuleHandleW(std::ptr::null());
-    if hinstance.is_null() {
-        tracing::warn!("set_window_icon: GetModuleHandleW returned null");
-        return;
-    }
-
-    // Load the big icon (32x32, for Alt+Tab / taskbar)
-    let icon_big = LoadImageW(
-        hinstance,
-        1 as *const u16, // Resource ID 1 (set by winres)
-        IMAGE_ICON,
-        32, 32,
-        LR_SHARED,
-    );
-    if !icon_big.is_null() {
-        SendMessageW(hwnd, WM_SETICON, ICON_BIG as usize, icon_big as isize);
-    }
-
-    // Load the small icon (16x16, for title bar)
-    let icon_small = LoadImageW(
-        hinstance,
-        1 as *const u16,
-        IMAGE_ICON,
-        16, 16,
-        LR_SHARED,
-    );
-    if !icon_small.is_null() {
-        SendMessageW(hwnd, WM_SETICON, ICON_SMALL as usize, icon_small as isize);
-    }
-
-    if !icon_big.is_null() || !icon_small.is_null() {
-        tracing::info!("Set window icon from embedded resource");
-    } else {
-        tracing::warn!("set_window_icon: no icon found in exe resource");
-    }
-}
-
-/// Synchronously tell the backend to close a window's workspace/tabs/shells.
-///
-/// Uses a raw TCP connection so no async runtime or extra crate is needed.
-/// Called from a background thread in `on_before_close` so the CEF UI thread
-/// is not blocked. Fire-and-forget: we write the request and don't read the response.
-fn backend_close_window(web_endpoint: &str, auth_key: &str, window_id: &str) {
-    use std::io::Write;
-
-    // Parse host:port from "http://127.0.0.1:PORT"
-    let addr_str = web_endpoint
-        .trim_start_matches("http://")
-        .trim_start_matches("https://");
-    let addr: std::net::SocketAddr = match addr_str.parse() {
-        Ok(a) => a,
-        Err(e) => {
-            tracing::warn!("[backend_close_window] cannot parse endpoint '{}': {}", web_endpoint, e);
-            return;
-        }
-    };
-
-    let body = serde_json::json!({
-        "service": "window",
-        "method": "CloseWindow",
-        "args": [window_id],
-        "uicontext": null,
-    }).to_string();
-    let request = format!(
-        "POST /agentmux/service?service=window&method=CloseWindow&authkey={} HTTP/1.1\r\n\
-         Host: 127.0.0.1\r\n\
-         Content-Type: application/json\r\n\
-         Content-Length: {}\r\n\
-         Connection: close\r\n\
-         \r\n\
-         {}",
-        auth_key, body.len(), body
-    );
-
-    dlog(&format!("backend_close_window: connecting to {} for window_id={}", addr, window_id));
-    let timeout = std::time::Duration::from_millis(2000);
-    match std::net::TcpStream::connect_timeout(&addr, timeout) {
-        Ok(mut stream) => {
-            stream.set_write_timeout(Some(timeout)).ok();
-            stream.set_read_timeout(Some(timeout)).ok();
-            match stream.write_all(request.as_bytes()) {
-                Ok(_) => {
-                    dlog(&format!("backend_close_window: sent request for window_id={}", window_id));
-                    // Read response to confirm the backend received it
-                    use std::io::Read;
-                    let mut resp = String::new();
-                    let _ = stream.read_to_string(&mut resp);
-                    let first_line = resp.lines().next().unwrap_or("(empty)").to_string();
-                    dlog(&format!("backend_close_window: response first line: {}", first_line));
-                }
-                Err(e) => dlog(&format!("backend_close_window: write failed: {}", e)),
-            }
-        }
-        Err(e) => dlog(&format!("backend_close_window: connect failed to {}: {}", addr, e)),
-    }
-}
